Add back Redis trigger
