+++ containerd-shim-spin/src/trigger.rs
@@ -1,11 +1,13 @@
-use std::collections::HashSet;
+use std::{collections::HashSet, future::Future, path::Path, pin::Pin};
 
 use anyhow::anyhow;
 use spin_app::locked::LockedApp;
 use spin_trigger::Trigger;
 // use spin_trigger::{loader, RuntimeConfig, TriggerExecutor, TriggerExecutorBuilder};
 use spin_trigger_http::HttpTrigger;
-// use spin_trigger_redis::RedisTrigger;
+use spin_trigger_redis::RedisTrigger;
+
+use crate::constants::{RUNTIME_CONFIG_PATH, SPIN_TRIGGER_WORKING_DIR};
 // use trigger_command::CommandTrigger;
 // use trigger_mqtt::MqttTrigger;
 // use trigger_sqs::SqsTrigger;
@@ -13,6 +15,39 @@ use spin_trigger_http::HttpTrigger;
 pub(crate) const HTTP_TRIGGER_TYPE: &str = <HttpTrigger as Trigger<
     <spin_runtime_factors::FactorsBuilder as spin_trigger::cli::RuntimeFactorsBuilder>::Factors,
 >>::TYPE;
+pub(crate) const REDIS_TRIGGER_TYPE: &str = <RedisTrigger as Trigger<
+    <spin_runtime_factors::FactorsBuilder as spin_trigger::cli::RuntimeFactorsBuilder>::Factors,
+>>::TYPE;
+
+/// Run the trigger with the given `App` and `ComponentLoader`.
+pub(crate) async fn run<
+    T: Trigger<<FactorsBuilder as RuntimeFactorsBuilder>::Factors> + 'static,
+>(
+    trigger: T,
+    app: App,
+    loader: &spin_trigger::loader::ComponentLoader,
+) -> anyhow::Result<Pin<Box<dyn Future<Output = anyhow::Result<()>>>>> {
+    let builder: TriggerAppBuilder<_, FactorsBuilder> = TriggerAppBuilder::new(trigger);
+
+    let future = builder
+        .run(app, factors_config(), Default::default(), loader)
+        .await?;
+    Ok(Box::pin(future))
+}
+
+/// Configuration for the factors.
+fn factors_config() -> FactorsConfig {
+    // Load in runtime config if one exists at expected location
+    let runtime_config_file = Path::new(RUNTIME_CONFIG_PATH)
+        .exists()
+        .then(|| RUNTIME_CONFIG_PATH.into());
+    let factors_config = FactorsConfig {
+        working_dir: SPIN_TRIGGER_WORKING_DIR.into(),
+        runtime_config_file,
+        ..Default::default()
+    };
+    factors_config
+}
 
 /// get the supported trigger types from the `LockedApp`.
 ///
@@ -29,8 +64,8 @@ pub(crate) const HTTP_TRIGGER_TYPE: &str = <HttpTrigger as Trigger<
 /// Note: this function returns a `HashSet` of supported trigger types. Duplicates are removed.
 pub(crate) fn get_supported_triggers(locked_app: &LockedApp) -> anyhow::Result<HashSet<String>> {
     let supported_triggers: HashSet<&str> = HashSet::from([
-        // RedisTrigger::TRIGGER_TYPE,
         HTTP_TRIGGER_TYPE,
+        REDIS_TRIGGER_TYPE,
         // SqsTrigger::TRIGGER_TYPE,
         // MqttTrigger::TRIGGER_TYPE,
         // CommandTrigger::TRIGGER_TYPE,
